Extract logic outside
